+++ app/src/main/java/com/example/httpurlconnection/services/MyAppURLs.java
@@ -1,4 +1,4 @@
-package com.example.httpurlconnection;
+package com.example.httpurlconnection.services;
 
 public interface MyAppURLs {
